UIEvent reporting refactoring: gui changes
